:recycle: refactor: Removendo funções declaradas muitas vezes
