+++ src/components/HistoricoFinanceiro.jsx
@@ -5,6 +5,7 @@ import Filtro from "../util/Filtro";
 import { Badge, Group, Pagination } from "@mantine/core";
 import { IconCash } from "@tabler/icons-react";
 import createAPI from "../services/createAPI";
+import {ArrumaHora} from "../util/ArrumaHora";
 
 const HistoricoFinanceiro = (id_usuario = 0) => {
   const [resposta, setResposta] = useState([]);
@@ -23,15 +24,6 @@ const HistoricoFinanceiro = (id_usuario = 0) => {
   const indexOfLastItem = currentPage * itemsPerPage;
   const indexOfFirstItem = indexOfLastItem - itemsPerPage;
   const currentItems = resposta.slice(indexOfFirstItem, indexOfLastItem);
-
-  function ArrumaHora(data) {
-    const data2 = data.split("T");
-    const data3 = data2[0].split("-");
-    const data4 = data3[2] + "/" + data3[1] + "/" + data3[0];
-    const data6 = data2[1].split(":");
-    const data5 = data4 + " " + (data6[0] - 3) + ":" + data6[1];
-    return data5;
-  }
 
   useEffect(() => {
     let id = id_usuario.id_usuario ?? 0;
