+++ src/components/ListarNotificacoesAgente.jsx
@@ -9,6 +9,7 @@ import createAPI from '../services/createAPI'
 import { Button, Group, Pagination } from '@mantine/core'
 import CarroLoading from './Carregamento'
 import { IconReload } from '@tabler/icons-react';
+import {ArrumaHora2, ArrumaHora3} from "../util/ArrumaHora";
 
 const ListarNotificacoesAgente = () => {
     const [data, setData] = useState([])
@@ -28,23 +29,6 @@ const ListarNotificacoesAgente = () => {
   const indexOfLastItem = currentPage * itemsPerPage;
   const indexOfFirstItem = indexOfLastItem - itemsPerPage;
   const currentItems = data.slice(indexOfFirstItem, indexOfLastItem);
-
-
-
-    function ArrumaHora(data, hora ) {
-        const data2 = data.split("T");
-        const data3 = data2[0].split("-");
-        const data4 = data3[2] + "/" + data3[1] + "/" + data3[0];
-        return data4;
-    }
-
-    function ArrumaHora2(data) {
-      const data2 = data.split("T");
-      const data6 = data2[1].split(":");
-      const data5 = (data6[0]-3) + ":" + data6[1] + ":";
-      const data7 = data5 + data6[2].split(".")[0];
-      return data7;
-    }
 
     useEffect(() => {
         localStorage.removeItem('autoInfracao')
@@ -96,7 +80,7 @@ const ListarNotificacoesAgente = () => {
         if (response.data.msg.resultado){
           setEstado(false)
           const newData = response.data.data.map((item) => ({
-            data: ArrumaHora(item.data),
+            data: ArrumaHora3(item.data),
             placa: item.veiculo.placa,
             cancelada: item.cancelada,
             cancelada_motivo: item.cancelada_motivo,
@@ -160,7 +144,7 @@ const ListarNotificacoesAgente = () => {
       setEstadoLoading(false)
       setEstado(false)
       const newData = response.data.data.map((item) => ({
-        data: ArrumaHora(item.data),
+        data: ArrumaHora3(item.data),
         placa: item.veiculo.placa,
         cancelada: item.cancelada,
         cancelada_motivo: item.cancelada_motivo,
