+++ src/components/RegistrarEstacionamentoParceiro.jsx
@@ -3,6 +3,7 @@ import React from "react";
 import { useState, useEffect } from "react";
 import VoltarComponente from "../util/VoltarComponente";
 import FuncTrocaComp from "../util/FuncTrocaComp";
+import validarPlaca from "../util/validarPlaca";
 import { useDisclosure } from "@mantine/hooks";
 import ModalPix from "./ModalPix";
 import { Button, Divider, Grid, Text } from "@mantine/core";
@@ -73,18 +74,6 @@ const RegistrarEstacionamentoParceiro = () => {
       handleRegistrar();
     }
   };
-
-  function validarPlaca(placa) {
-    placa = placa.replace(/\s+/g, '');
-    const regexPlacaAntiga = /^[a-zA-Z]{3}\d{4}$/;
-    const regexPlacaNova = /^[A-Z]{3}[0-9][A-Z][0-9]{2}$/;
-
-    if (regexPlacaAntiga.test(placa) || regexPlacaNova.test(placa)) {
-      return true;
-    } else {
-      return false;
-    }
-  }
 
   const handlePlaca = () => {
     const clicado = document.getElementById("flexSwitchCheckDefault").checked;
