+++ src/util/ImpressaoFecharCaixa.jsx
@@ -1,17 +1,10 @@
-const ImpressaoFecharCaixa = async (item, valor, monitor) => {
-  const FormatDate = (date) => {
-    const data = new Date(date);
-    const year = data.getFullYear();
-    const month = String(data.getMonth() + 1).padStart(2, "0");
-    const day = String(data.getDate()).padStart(2, "0");
-    const formattedDate = `${day}/${month}-${year}`;
+import {FormatDateBr} from "../util/formatDate";
 
-    return formattedDate;
-  };
+const ImpressaoFecharCaixa = async (item, valor, monitor) => {
 
   const json = {
     tipo: "FECHAMENTO DE CAIXA",
-    dataEmissao: FormatDate(item.data),
+    dataEmissao: FormatDateBr(item.data),
     monitor: monitor,
     valorMovimento: item.valor_movimentos,
     valorAbertura: item.valor_abertura,
