+++ src/util/ImpressaoTicketEstacionamento.jsx
@@ -1,5 +1,5 @@
 import axios from "axios";
-
+import calcularValidade from "../util/CalcularValidade";
 const ImpressaoTicketEstacionamento = async (
   via,
   tempoChegada,
@@ -11,18 +11,6 @@ const ImpressaoTicketEstacionamento = async (
   tempoValor,
   notificacao
 ) => {
-  const calcularValidade = (horaInicio, duracao) => {
-    const [horas, minutos, segundos] = duracao.split(":").map(Number);
-    const dataInicio = new Date(`2000-01-01T${horaInicio}`);
-    const dataValidade = new Date(
-      dataInicio.getTime() + horas * 3600000 + minutos * 60000 + segundos * 1000
-    );
-    const horaValidade = dataValidade.toLocaleTimeString("pt-BR", {
-      timeZone: "America/Sao_Paulo",
-    });
-    return horaValidade;
-  };
-
   const duracao = tempo;
   let horaValidade = calcularValidade(tempoChegada, duracao);
