+++ src/util/extrairNumeros.jsx
@@ -1,0 +1,4 @@
+function extrairNumeros(string) {
+    return string ? string.replace(/\D/g, "") : string;
+  }
+  export default extrairNumeros;
